float change

+++ nn_train_FNO.py
@@ -75,9 +75,9 @@ for ep in range(0, epochs+1):
     for step in range(0,trainN,batch_size):
         indices = np.random.permutation(np.arange(start=step, step=1,stop=step+batch_size))
         input_batch, label_batch, du_label_batch = input_train_torch[indices], label_train_torch[indices], du_label_torch[indices]
-        input_batch = torch.reshape(input_batch,(batch_size,input_size,1))
-        label_batch = torch.reshape(label_batch,(batch_size,input_size,1))
-        du_label_batch = torch.reshape(du_label_batch,(batch_size,input_size,1))
+        input_batch = torch.reshape(input_batch,(batch_size,input_size,1)).float()
+        label_batch = torch.reshape(label_batch,(batch_size,input_size,1)).float()
+        du_label_batch = torch.reshape(du_label_batch,(batch_size,input_size,1)).float()
 
         #pick a random boundary batch
         optimizer.zero_grad()
